+++ package.json_bak
@@ -4,9 +4,6 @@
   "private": true,
   "homepage": "https://valise.stegra.ca",
   "dependencies": {
-    "@babel/plugin-transform-dynamic-import": "^7.23.4",
-    "@babel/plugin-transform-json-strings": "^7.23.4",
-    "@babel/plugin-transform-unicode-property-regex": "^7.23.3",
     "@jridgewell/sourcemap-codec": "^1.4.15",
     "@react-spring/web": "^9.7.2",
     "@rollup/plugin-terser": "^0.4.3",
@@ -17,13 +14,11 @@
     "animate-css-grid": "^1.5.1",
     "bootstrap": "^5.2.3",
     "cheerio": "^1.0.0-rc.12",
-    "color-image-detector": "^1.0.3",
     "eslint-config-react-app": "^7.0.1",
     "extract-colors": "^4.0.2",
     "get-image-colors": "^4.0.1",
     "gh-pages": "^5.0.0",
     "holderjs": "^2.9.9",
-    "ini": "^4.1.2",
     "parse5-htmlparser2-tree-adapter": "^7.0.0",
     "path-browserify": "^1.0.1",
     "react": "^18.2.0",
@@ -32,7 +27,8 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-modal-image": "^2.6.0",
-    "react-scripts": "5.0.0",
+    // "react-palette": "^1.0.2",
+    "react-scripts": "^5.0.1",
     "react-tabs": "^6.0.2",
     "sass": "^1.62.1",
     "text-to-svg": "^3.1.5",
