+++ src/components/biography/Biography.jsx
@@ -1,5 +1,4 @@
 import React from "react";
-import {animated, useSpring } from "@react-spring/web";
 import ProfilePic from "./images/profile-photo-small.jpg";
 import "./biography.css";
